requirements.txt
numpy
pandas
torch
tifffile
pillow
python-dotenv
pydantic
scikit-image
pathlib
opencv-python